refactor(isFeatureFlagEnabled): move to .server file
